+++ weblate/trans/signals.py
@@ -23,7 +23,7 @@ from django.dispatch import Signal
 
 vcs_post_push = Signal(providing_args=['component'])
 vcs_post_update = Signal(providing_args=['component', 'previous_head'])
-vcs_pre_commit = Signal(providing_args=['translation'])
+vcs_pre_commit = Signal(providing_args=['translation', 'author'])
 vcs_post_commit = Signal(providing_args=['translation'])
 translation_post_add = Signal(providing_args=['translation'])
 unit_pre_create = Signal(providing_args=['unit'])