+++ docs/changes.rst
@@ -10,6 +10,7 @@ Released on ? 2018.
 * New addon for automatic component discovery.
 * The import_project management command has now slightly different parameters.
 * Added basic support for Windows RC files.
+* New addon to store contributor names in PO file headers.
 
 weblate 2.20
 ------------